create xml node if not exist
