+++ EMEProToolKit/EMEProToolkitSrc/EMEProToolkit.csproj
@@ -1036,6 +1036,7 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TestButton1.cs" />
+    <Compile Include="ThumbnailUpdater.cs" />
     <Compile Include="Utils\LanguageConverter.cs" />
     <Compile Include="Utils\MetadataException.cs" />
     <Compile Include="Utils\MetadataStyle.cs" />
